Another loop another debugging session.

+++ src/core/ee/interpreter/interpreter_cop0.cpp
@@ -3,7 +3,7 @@
 
 void ee_interpreter::mfc0(ee_inst inst)
 {
-    rGPR[inst.RT].ud[0] = (s64)(s32)rCOP0[inst.RD];
+    rGPR[inst.RT].ud[0] = rCOP0[inst.RD];
 }
 
 void ee_interpreter::mtc0(ee_inst inst)
